Migrate safe-delete to TypeScript Plugin SDK
