+++ plugins/safe-delete/README.md
@@ -1,8 +1,8 @@
 # Safe Delete Native Plugin
 
-`safe_delete` is a dependency-free WebCodex Native Tool Plugin that moves one
-ordinary file or directory to the operating system Trash/Recycle Bin instead of
-permanently deleting it.
+`safe_delete` is a TypeScript-authored WebCodex Native Tool Plugin built with
+`@yyjeqhc/webcodex-plugin-sdk`. It moves one ordinary file or directory to the
+operating system Trash/Recycle Bin instead of permanently deleting it.
 
 It is intentionally a Plugin rather than a built-in WebCodex filesystem tool.
 Install it only on Runners where you want the model to have this additional
@@ -46,8 +46,10 @@ unlink; it falls through to another Trash backend instead.
 ## Runner configuration
 
 Set `cwd` to the exact project/root you want this Plugin to be allowed to trash
-from. Use an absolute path to the Plugin script when the Plugin code lives
-outside that root:
+from. The checked-in `plugin.mjs` is a stable Runner entrypoint that loads the
+compiled TypeScript output, so build the Plugin before check/reload. Use an
+absolute path to that entrypoint when the Plugin code lives outside the authority
+root:
 
 ```toml
 [plugins]
@@ -85,9 +87,19 @@ Possible structured outcomes are `trashed`, `already_absent`, `rejected`,
 
 ## Development
 
-No npm install is required:
+The SDK is consumed from this repository through a local `file:` dependency; no
+workspace/package-manager layer is required. From the repository root:
 
 ```bash
-node --check plugins/safe-delete/plugin.mjs
-node --test plugins/safe-delete/plugin.test.mjs
+npm ci --prefix npm/plugin-sdk
+npm --prefix npm/plugin-sdk run build
+npm ci --prefix plugins/safe-delete
+npm --prefix plugins/safe-delete run typecheck
+npm --prefix plugins/safe-delete test
 ```
+
+The TypeScript entrypoint owns only authoring/runtime boilerplate. Path authority,
+Trash backend selection, and result classification remain in the safe-delete
+domain implementation, while the Rust Runner remains authoritative for Plugin
+admission, schema/runtime validation, timeout, process lifecycle, and
+`OutcomeUnknown` handling.
