+++ plugins/safe-delete/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@yyjeqhc/webcodex-safe-delete-plugin",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "build": "tsc -p tsconfig.json",
+    "test": "npm run build --silent && node --test plugin.test.mjs"
+  },
+  "dependencies": {
+    "@yyjeqhc/webcodex-plugin-sdk": "file:../../npm/plugin-sdk"
+  },
+  "devDependencies": {
+    "@types/node": "^18.19.0",
+    "typescript": "^5.4.0"
+  }
+}
